:bug:  Fix file hiearchy for statistiques pages
webpack did not see them when building js bundles
