+++ src/views/pages/index.ts
@@ -1,7 +1,9 @@
-export * from './modificationRequestPage'
-export * from './garantiesFinancieresPage'
-export * from './newModificationRequestPage'
-export * from './projectDetailsPage'
-export * from './statistiques'
-export * from './Error'
-export * from './homePage'
+export * from './garantiesFinancieresPage';
+export * from './homePage';
+export * from './modificationRequestPage';
+export * from './newModificationRequestPage';
+export * from './projectDetailsPage';
+export * from './AdemeStatistiquesPage';
+export * from './AdminStatistiquesPage';
+export * from './Error';
+export * from './StatistiquesPage';
