--- src/views/pages/statistiques/Statistiques.stories.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import { Statistiques } from './Statistiques'
-import makeFakeRequest from '../../../__tests__/fixtures/request'
-
-export default { title: 'Statistiques' }
-
-const props = {
-  request: makeFakeRequest(),
-  mainIframeUrl: '',
-  mapIframeUrl: '',
-}
-
-export const empty = () => <Statistiques {...props} />
