--- src/views/pages/statistiques/index.ts
@@ -1,3 +1,0 @@
-export * from './AdemeStatistiques';
-export * from './AdminStatistiques';
-export * from './Statistiques';
